requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
aiohttp>=3.9.0
asyncio>=3.4.3
python-socks>=2.4.3
async_timeout>=4.0.3